+++ inst/include/jsonify/to_json/writers/complex.hpp
@@ -329,8 +329,9 @@ namespace complex {
           }
           
           // LIST NAMES
-          Rcpp::IntegerVector int_names = Rcpp::seq(1, lst.size());
-          Rcpp::CharacterVector list_names = Rcpp::as< Rcpp::CharacterVector >( int_names );
+          //Rcpp::IntegerVector int_names = Rcpp::seq(1, lst.size());
+          //Rcpp::CharacterVector list_names = Rcpp::as< Rcpp::CharacterVector >( int_names );
+          Rcpp::CharacterVector list_names( lst.size() );
           has_names = lst.hasAttribute("names");
           
           if ( has_names ) {