+++ tests/testthat/test-to_json_SEXPTYPE.R
@@ -6,7 +6,7 @@ test_that("SEXPTYPES are convertedt to JSON", {
   
   ## closure & language
   js <- to_json( to_json, unbox = TRUE )
-  expect_equal( as.character( js ), '{"x":"","unbox":false,"digits":{},"numeric_dates":true,"factors_as_string":true,"by":"row","7":["{",["if",["%in%","col","by"],["<-","by","column"]],["<-","by",{"1":"match.arg","2":"by","choices":["c","row","column"]}],["<-","digits",["handle_digits","digits"]],["rcpp_to_json","x","unbox","digits","numeric_dates","factors_as_string","by"]]}')
+  expect_equal( as.character( js ), '{"x":"","unbox":false,"digits":{},"numeric_dates":true,"factors_as_string":true,"by":"row","":["{",["if",["%in%","col","by"],["<-","by","column"]],["<-","by",{"":"match.arg","":"by","choices":["c","row","column"]}],["<-","digits",["handle_digits","digits"]],["rcpp_to_json","x","unbox","digits","numeric_dates","factors_as_string","by"]]}')
   expect_true( validate_json( js ) ) 
   
   ## builtinsxp
@@ -21,7 +21,7 @@ test_that("SEXPTYPES are convertedt to JSON", {
   
   ## clouse & language
   js <- to_json( print, unbox = TRUE )
-  expect_equal( as.character( js ), '{"x":"","...":"","3":["UseMethod","print"]}')
+  expect_equal( as.character( js ), '{"x":"","...":"","":["UseMethod","print"]}')
   expect_true( validate_json( js ) ) 
   
   ## envsxp