+++ tests/testthat/test-to_json_lists.R
@@ -8,7 +8,7 @@ test_that("empty lists are handled", {
 
 test_that("list structures jsonified", {
   lst <- list(x = 1, y = c("a","b","c"), z = list("hello", b = "world"))
-  expect_equal( as.character(to_json( lst )), "{\"x\":[1.0],\"y\":[\"a\",\"b\",\"c\"],\"z\":{\"1\":[\"hello\"],\"b\":[\"world\"]}}")
+  expect_equal( as.character(to_json( lst )), "{\"x\":[1.0],\"y\":[\"a\",\"b\",\"c\"],\"z\":{\"\":[\"hello\"],\"b\":[\"world\"]}}")
   expect_true( validate_json( to_json(lst) ) ) 
 })
 
@@ -21,7 +21,7 @@ test_that("empty lists are handled and unboxed", {
 
 test_that("list structures jsonified and unboxed", {
   lst <- list(x = 1, y = c("a","b","c"), z = list("hello", b = "world"))
-  expect_equal( as.character(to_json( lst, unbox = T )), "{\"x\":1.0,\"y\":[\"a\",\"b\",\"c\"],\"z\":{\"1\":\"hello\",\"b\":\"world\"}}")
+  expect_equal( as.character(to_json( lst, unbox = T )), "{\"x\":1.0,\"y\":[\"a\",\"b\",\"c\"],\"z\":{\"\":\"hello\",\"b\":\"world\"}}")
   expect_true( validate_json( to_json( lst, unbox = T) ) ) 
 })
 
@@ -66,7 +66,7 @@ test_that("matrices inside lists are correctly converted", {
   l <- list( x = m1, y = m2, m3 )
   js <- to_json( l )
   expect_true( validate_json( js ) )
-  expected <- '{"x":[[1,3],[2,4]],"y":[["a","c","e"],["b","d","f"]],"3":[[1.0]]}'
+  expected <- '{"x":[[1,3],[2,4]],"y":[["a","c","e"],["b","d","f"]],"":[[1.0]]}'
   expect_equal( as.character( js ), expected )
   
   m1 <- matrix(1:4, ncol = 2)
@@ -75,7 +75,7 @@ test_that("matrices inside lists are correctly converted", {
   l <- list( x = m1, y = m2, m3 )
   js <- to_json( l, unbox = TRUE )
   expect_true( validate_json( js ) )
-  expected <- '{"x":[[1,3],[2,4]],"y":[["a","c","e"],["b","d","f"]],"3":[1.0]}'
+  expected <- '{"x":[[1,3],[2,4]],"y":[["a","c","e"],["b","d","f"]],"":[1.0]}'
   expect_equal( as.character( js ), expected )
   
 })
@@ -101,3 +101,4 @@ test_that("factors in lists become characters", {
   l2 <- list( x = 1:3 )
   expect_true( js == to_json( l2 ) )
 })
+